[runtime] Move basic pe_file functionality into proclib
