+++ mono/metadata/w32process-unix.c
@@ -253,245 +253,6 @@ typedef struct {
 #define VS_FFI_STRUCVERSION	0x00010000
 #endif
 
-#define IMAGE_NUMBEROF_DIRECTORY_ENTRIES 16
-
-#define IMAGE_DIRECTORY_ENTRY_EXPORT	0
-#define IMAGE_DIRECTORY_ENTRY_IMPORT	1
-#define IMAGE_DIRECTORY_ENTRY_RESOURCE	2
-
-#define IMAGE_SIZEOF_SHORT_NAME	8
-
-#if G_BYTE_ORDER != G_LITTLE_ENDIAN
-#define IMAGE_DOS_SIGNATURE	0x4d5a
-#define IMAGE_NT_SIGNATURE	0x50450000
-#define IMAGE_NT_OPTIONAL_HDR32_MAGIC	0xb10
-#define IMAGE_NT_OPTIONAL_HDR64_MAGIC	0xb20
-#else
-#define IMAGE_DOS_SIGNATURE	0x5a4d
-#define IMAGE_NT_SIGNATURE	0x00004550
-#define IMAGE_NT_OPTIONAL_HDR32_MAGIC	0x10b
-#define IMAGE_NT_OPTIONAL_HDR64_MAGIC	0x20b
-#endif
-
-typedef struct {
-	guint16 e_magic;
-	guint16 e_cblp;
-	guint16 e_cp;
-	guint16 e_crlc;
-	guint16 e_cparhdr;
-	guint16 e_minalloc;
-	guint16 e_maxalloc;
-	guint16 e_ss;
-	guint16 e_sp;
-	guint16 e_csum;
-	guint16 e_ip;
-	guint16 e_cs;
-	guint16 e_lfarlc;
-	guint16 e_ovno;
-	guint16 e_res[4];
-	guint16 e_oemid;
-	guint16 e_oeminfo;
-	guint16 e_res2[10];
-	guint32 e_lfanew;
-} IMAGE_DOS_HEADER;
-
-typedef struct {
-	guint16 Machine;
-	guint16 NumberOfSections;
-	guint32 TimeDateStamp;
-	guint32 PointerToSymbolTable;
-	guint32 NumberOfSymbols;
-	guint16 SizeOfOptionalHeader;
-	guint16 Characteristics;
-} IMAGE_FILE_HEADER;
-
-typedef struct {
-	guint32 VirtualAddress;
-	guint32 Size;
-} IMAGE_DATA_DIRECTORY;
-
-typedef struct {
-	guint16 Magic;
-	guint8 MajorLinkerVersion;
-	guint8 MinorLinkerVersion;
-	guint32 SizeOfCode;
-	guint32 SizeOfInitializedData;
-	guint32 SizeOfUninitializedData;
-	guint32 AddressOfEntryPoint;
-	guint32 BaseOfCode;
-	guint32 BaseOfData;
-	guint32 ImageBase;
-	guint32 SectionAlignment;
-	guint32 FileAlignment;
-	guint16 MajorOperatingSystemVersion;
-	guint16 MinorOperatingSystemVersion;
-	guint16 MajorImageVersion;
-	guint16 MinorImageVersion;
-	guint16 MajorSubsystemVersion;
-	guint16 MinorSubsystemVersion;
-	guint32 Win32VersionValue;
-	guint32 SizeOfImage;
-	guint32 SizeOfHeaders;
-	guint32 CheckSum;
-	guint16 Subsystem;
-	guint16 DllCharacteristics;
-	guint32 SizeOfStackReserve;
-	guint32 SizeOfStackCommit;
-	guint32 SizeOfHeapReserve;
-	guint32 SizeOfHeapCommit;
-	guint32 LoaderFlags;
-	guint32 NumberOfRvaAndSizes;
-	IMAGE_DATA_DIRECTORY DataDirectory[IMAGE_NUMBEROF_DIRECTORY_ENTRIES];
-} IMAGE_OPTIONAL_HEADER32;
-
-typedef struct {
-	guint16 Magic;
-	guint8 MajorLinkerVersion;
-	guint8 MinorLinkerVersion;
-	guint32 SizeOfCode;
-	guint32 SizeOfInitializedData;
-	guint32 SizeOfUninitializedData;
-	guint32 AddressOfEntryPoint;
-	guint32 BaseOfCode;
-	guint64 ImageBase;
-	guint32 SectionAlignment;
-	guint32 FileAlignment;
-	guint16 MajorOperatingSystemVersion;
-	guint16 MinorOperatingSystemVersion;
-	guint16 MajorImageVersion;
-	guint16 MinorImageVersion;
-	guint16 MajorSubsystemVersion;
-	guint16 MinorSubsystemVersion;
-	guint32 Win32VersionValue;
-	guint32 SizeOfImage;
-	guint32 SizeOfHeaders;
-	guint32 CheckSum;
-	guint16 Subsystem;
-	guint16 DllCharacteristics;
-	guint64 SizeOfStackReserve;
-	guint64 SizeOfStackCommit;
-	guint64 SizeOfHeapReserve;
-	guint64 SizeOfHeapCommit;
-	guint32 LoaderFlags;
-	guint32 NumberOfRvaAndSizes;
-	IMAGE_DATA_DIRECTORY DataDirectory[IMAGE_NUMBEROF_DIRECTORY_ENTRIES];
-} IMAGE_OPTIONAL_HEADER64;
-
-#if SIZEOF_VOID_P == 8
-typedef IMAGE_OPTIONAL_HEADER64 IMAGE_OPTIONAL_HEADER;
-#else
-typedef IMAGE_OPTIONAL_HEADER32 IMAGE_OPTIONAL_HEADER;
-#endif
-
-typedef struct {
-	guint32 Signature;
-	IMAGE_FILE_HEADER FileHeader;
-	IMAGE_OPTIONAL_HEADER32 OptionalHeader;
-} IMAGE_NT_HEADERS32;
-
-typedef struct {
-	guint32 Signature;
-	IMAGE_FILE_HEADER FileHeader;
-	IMAGE_OPTIONAL_HEADER64 OptionalHeader;
-} IMAGE_NT_HEADERS64;
-
-#if SIZEOF_VOID_P == 8
-typedef IMAGE_NT_HEADERS64 IMAGE_NT_HEADERS;
-#else
-typedef IMAGE_NT_HEADERS32 IMAGE_NT_HEADERS;
-#endif
-
-typedef struct {
-	guint8 Name[IMAGE_SIZEOF_SHORT_NAME];
-	union {
-		guint32 PhysicalAddress;
-		guint32 VirtualSize;
-	} Misc;
-	guint32 VirtualAddress;
-	guint32 SizeOfRawData;
-	guint32 PointerToRawData;
-	guint32 PointerToRelocations;
-	guint32 PointerToLinenumbers;
-	guint16 NumberOfRelocations;
-	guint16 NumberOfLinenumbers;
-	guint32 Characteristics;
-} IMAGE_SECTION_HEADER;
-
-#define IMAGE_FIRST_SECTION32(header) ((IMAGE_SECTION_HEADER *)((gsize)(header) + G_STRUCT_OFFSET (IMAGE_NT_HEADERS32, OptionalHeader) + GUINT16_FROM_LE (((IMAGE_NT_HEADERS32 *)(header))->FileHeader.SizeOfOptionalHeader)))
-
-#define RT_CURSOR	0x01
-#define RT_BITMAP	0x02
-#define RT_ICON		0x03
-#define RT_MENU		0x04
-#define RT_DIALOG	0x05
-#define RT_STRING	0x06
-#define RT_FONTDIR	0x07
-#define RT_FONT		0x08
-#define RT_ACCELERATOR	0x09
-#define RT_RCDATA	0x0a
-#define RT_MESSAGETABLE	0x0b
-#define RT_GROUP_CURSOR	0x0c
-#define RT_GROUP_ICON	0x0e
-#define RT_VERSION	0x10
-#define RT_DLGINCLUDE	0x11
-#define RT_PLUGPLAY	0x13
-#define RT_VXD		0x14
-#define RT_ANICURSOR	0x15
-#define RT_ANIICON	0x16
-#define RT_HTML		0x17
-#define RT_MANIFEST	0x18
-
-typedef struct {
-	guint32 Characteristics;
-	guint32 TimeDateStamp;
-	guint16 MajorVersion;
-	guint16 MinorVersion;
-	guint16 NumberOfNamedEntries;
-	guint16 NumberOfIdEntries;
-} IMAGE_RESOURCE_DIRECTORY;
-
-typedef struct {
-	union {
-		struct {
-#if G_BYTE_ORDER == G_BIG_ENDIAN
-			guint32 NameIsString:1;
-			guint32 NameOffset:31;
-#else
-			guint32 NameOffset:31;
-			guint32 NameIsString:1;
-#endif
-		};
-		guint32 Name;
-#if G_BYTE_ORDER == G_BIG_ENDIAN
-		struct {
-			guint16 __wapi_big_endian_padding;
-			guint16 Id;
-		};
-#else
-		guint16 Id;
-#endif
-	};
-	union {
-		guint32 OffsetToData;
-		struct {
-#if G_BYTE_ORDER == G_BIG_ENDIAN
-			guint32 DataIsDirectory:1;
-			guint32 OffsetToDirectory:31;
-#else
-			guint32 OffsetToDirectory:31;
-			guint32 DataIsDirectory:1;
-#endif
-		};
-	};
-} IMAGE_RESOURCE_DIRECTORY_ENTRY;
-
-typedef struct {
-	guint32 OffsetToData;
-	guint32 Size;
-	guint32 CodePage;
-	guint32 Reserved;
-} IMAGE_RESOURCE_DATA_ENTRY;
-
 #define VOS_UNKNOWN		0x00000000
 #define VOS_DOS			0x00010000
 #define VOS_OS216		0x00020000
@@ -3140,91 +2901,6 @@ find_pe_file_resources (gpointer file_map, guint32 map_size, guint32 res_id, gui
 	}
 }
 
-static gpointer
-map_pe_file (gunichar2 *filename, gint32 *map_size, void **handle)
-{
-	gchar *filename_ext = NULL;
-	gchar *located_filename = NULL;
-	int fd = -1;
-	struct stat statbuf;
-	gpointer file_map = NULL;
-
-	/* According to the MSDN docs, a search path is applied to
-	 * filename.  FIXME: implement this, for now just pass it
-	 * straight to open
-	 */
-
-	filename_ext = mono_unicode_to_external (filename);
-	if (filename_ext == NULL) {
-		mono_trace (G_LOG_LEVEL_DEBUG, MONO_TRACE_IO_LAYER_PROCESS, "%s: unicode conversion returned NULL", __func__);
-
-		mono_w32error_set_last (ERROR_INVALID_NAME);
-		goto exit;
-	}
-
-	fd = open (filename_ext, O_RDONLY, 0);
-	if (fd == -1 && (errno == ENOENT || errno == ENOTDIR) && IS_PORTABILITY_SET) {
-		gint saved_errno = errno;
-
-		located_filename = mono_portability_find_file (filename_ext, TRUE);
-		if (!located_filename) {
-			errno = saved_errno;
-
-			mono_trace (G_LOG_LEVEL_DEBUG, MONO_TRACE_IO_LAYER_PROCESS, "%s: Error opening file %s (1): %s", __func__, filename_ext, strerror (errno));
-			goto error;
-		}
-
-		fd = open (located_filename, O_RDONLY, 0);
-		if (fd == -1) {
-			mono_trace (G_LOG_LEVEL_DEBUG, MONO_TRACE_IO_LAYER_PROCESS, "%s: Error opening file %s (2): %s", __func__, filename_ext, strerror (errno));
-			goto error;
-		}
-	}
-	else if (fd == -1) {
-		mono_trace (G_LOG_LEVEL_DEBUG, MONO_TRACE_IO_LAYER_PROCESS, "%s: Error opening file %s (3): %s", __func__, filename_ext, strerror (errno));
-		goto error;
-	}
-
-	if (fstat (fd, &statbuf) == -1) {
-		mono_trace (G_LOG_LEVEL_DEBUG, MONO_TRACE_IO_LAYER_PROCESS, "%s: Error stat()ing file %s: %s", __func__, filename_ext, strerror (errno));
-		goto error;
-	}
-	*map_size = statbuf.st_size;
-
-	/* Check basic file size */
-	if (statbuf.st_size < sizeof(IMAGE_DOS_HEADER)) {
-		mono_trace (G_LOG_LEVEL_DEBUG, MONO_TRACE_IO_LAYER_PROCESS, "%s: File %s is too small: %lld", __func__, filename_ext, (long long) statbuf.st_size);
-
-		mono_w32error_set_last (ERROR_BAD_LENGTH);
-		goto exit;
-	}
-
-	file_map = mono_file_map (statbuf.st_size, MONO_MMAP_READ | MONO_MMAP_PRIVATE, fd, 0, handle);
-	if (file_map == NULL) {
-		mono_trace (G_LOG_LEVEL_DEBUG, MONO_TRACE_IO_LAYER_PROCESS, "%s: Error mmap()int file %s: %s", __func__, filename_ext, strerror (errno));
-		goto error;
-	}
-exit:
-	if (fd != -1)
-		close (fd);
-	g_free (located_filename);
-	g_free (filename_ext);
-	return file_map;
-error:
-	mono_w32error_set_last (mono_w32error_unix_to_win32 (errno));
-	goto exit;
-}
-
-static void
-unmap_pe_file (gpointer file_map, void *handle)
-{
-	gint res;
-
-	res = mono_file_unmap (file_map, handle);
-	if (G_UNLIKELY (res != 0))
-		g_error ("%s: mono_file_unmap failed, error: \"%s\" (%d)", __func__, g_strerror (errno), errno);
-}
-
 static guint32
 unicode_chars (const gunichar2 *str)
 {
@@ -3645,38 +3321,6 @@ big_up (gconstpointer datablock, guint32 size)
 #endif
 
 gboolean
-mono_w32process_time_date_stamp (gunichar2 *filename, guint32 *out)
-{
-	void *map_handle;
-	gint32 map_size;
-	gpointer file_map = map_pe_file (filename, &map_size, &map_handle);
-	if (!file_map)
-		return FALSE;
-
-	/* Figure this out when we support 64bit PE files */
-	if (1) {
-		IMAGE_DOS_HEADER *dos_header = (IMAGE_DOS_HEADER *)file_map;
-		if (dos_header->e_magic != IMAGE_DOS_SIGNATURE) {
-			unmap_pe_file (file_map, map_handle);
-			return FALSE;
-		}
-
-		IMAGE_NT_HEADERS32 *nt_headers = (IMAGE_NT_HEADERS32 *)((guint8 *)file_map + GUINT32_FROM_LE (dos_header->e_lfanew));
-		if (nt_headers->Signature != IMAGE_NT_SIGNATURE) {
-			unmap_pe_file (file_map, map_handle);
-			return FALSE;
-		}
-
-		*out = nt_headers->FileHeader.TimeDateStamp;
-	} else {
-		g_assert_not_reached ();
-	}
-
-	unmap_pe_file (file_map, map_handle);
-	return TRUE;
-}
-
-gboolean
 mono_w32process_get_fileversion_info (gunichar2 *filename, gpointer *data)
 {
 	gpointer file_map;
@@ -3688,13 +3332,13 @@ mono_w32process_get_fileversion_info (gunichar2 *filename, gpointer *data)
 	g_assert (data);
 	*data = NULL;
 
-	file_map = map_pe_file (filename, &map_size, &map_handle);
+	file_map = mono_pe_file_map (filename, &map_size, &map_handle);
 	if (!file_map)
 		return FALSE;
 
 	versioninfo = find_pe_file_resources (file_map, map_size, RT_VERSION, 0, &datasize);
 	if (!versioninfo) {
-		unmap_pe_file (file_map, map_handle);
+		mono_pe_file_unmap (file_map, map_handle);
 		return FALSE;
 	}
 
@@ -3708,7 +3352,7 @@ mono_w32process_get_fileversion_info (gunichar2 *filename, gpointer *data)
 	big_up (*data, datasize);
 #endif
 
-	unmap_pe_file (file_map, map_handle);
+	mono_pe_file_unmap (file_map, map_handle);
 
 	return TRUE;
 }
